fix circular dependency and isolate survey access checks
