+++ lms/djangoapps/courseware/access_response.py
@@ -216,17 +216,6 @@ class NoAllowedPartitionGroupsError(AccessError):
         super(NoAllowedPartitionGroupsError, self).__init__(error_code, developer_message, user_message)
 
 
-class SurveyRequiredAccessError(AccessError):
-    """
-    Access denied because the user has not completed a required survey
-    """
-    def __init__(self):
-        error_code = "survey_required"
-        developer_message = u"User must complete a survey"
-        user_message = _(u"You must complete a survey")
-        super(SurveyRequiredAccessError, self).__init__(error_code, developer_message, user_message)
-
-
 class EnrollmentRequiredAccessError(AccessError):
     """
     Access denied because the user must be enrolled in the course
