+++ lms/djangoapps/survey/utils.py
@@ -4,7 +4,19 @@ Utilities for determining whether or not a survey needs to be completed.
 
 
 from lms.djangoapps.courseware.access import has_access
+from lms.djangoapps.courseware.access_response import AccessError, ACCESS_GRANTED
 from survey.models import SurveyAnswer, SurveyForm
+
+
+class SurveyRequiredAccessError(AccessError):
+    """
+    Access denied because the user has not completed a required survey
+    """
+    def __init__(self):
+        error_code = "survey_required"
+        developer_message = u"User must complete a survey"
+        user_message = _(u"You must complete a survey")
+        super(SurveyRequiredAccessError, self).__init__(error_code, developer_message, user_message)
 
 
 def is_survey_required_for_course(course_descriptor):
@@ -14,7 +26,7 @@ def is_survey_required_for_course(course_descriptor):
 
     # Check to see that the survey is required in the CourseDescriptor.
     if not getattr(course_descriptor, 'course_survey_required', False):
-        return False
+        return SurveyRequiredAccessError()
 
     # Check that the specified Survey for the course exists.
     return SurveyForm.get(course_descriptor.course_survey_name, throw_if_not_found=False)
@@ -22,23 +34,26 @@ def is_survey_required_for_course(course_descriptor):
 
 def is_survey_required_and_unanswered(user, course_descriptor):
     """
-    Returns whether a user is required to answer the survey and has yet to do so.
+    Checks whether a user is required to answer the survey and has yet to do so.
+
+    Returns:
+        AccessResponse: Either ACCESS_GRANTED or SurveyRequiredAccessError.
     """
 
     if not is_survey_required_for_course(course_descriptor):
-        return False
+        return SurveyRequiredAccessError()
 
     # anonymous users do not need to answer the survey
     if user.is_anonymous:
-        return False
+        return SurveyRequiredAccessError()
 
     # course staff do not need to answer survey
     has_staff_access = has_access(user, 'staff', course_descriptor)
     if has_staff_access:
-        return False
+        return SurveyRequiredAccessError()
 
     # survey is required and it exists, let's see if user has answered the survey
     survey = SurveyForm.get(course_descriptor.course_survey_name)
     answered_survey = SurveyAnswer.do_survey_answers_exist(survey, user)
     if not answered_survey:
-        return True
+        return ACCESS_GRANTED
